Removed streams from Collections overrides
It seems weird to rely on streams in a fairly bare Collection
implementation. They should be something that should be used at a higher
level, if the treap were actually used, but not here. They were also
taking away shortcutting in `containsAll(Collection<?> c)`, making it
less efficient for no reason.